refactor: 도메인 공통 서비스 패키지 구조 변경
