+++ src/main/java/com/gomo/app/interest/application/OrderUpdateMajorInterestUseCase.java
@@ -3,7 +3,7 @@ package com.gomo.app.interest.application;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.gomo.app.common.application.ApplicationService;
-import com.gomo.app.common.domain.OrderChanger;
+import com.gomo.app.common.domain.service.OrderChanger;
 import com.gomo.app.interest.domain.repository.MajorInterestRepository;
 import com.gomo.app.interest.presentation.request.OrderUpdateMajorInterestRequest;
 import com.gomo.app.member.domain.model.MemberId;
