+++ src/main/java/com/gomo/app/interest/domain/model/MajorInterest.java
@@ -4,8 +4,8 @@ import java.util.UUID;
 
 import com.gomo.app.common.domain.Authorizable;
 import com.gomo.app.common.domain.BaseAudit;
-import com.gomo.app.common.domain.DisplayOrder;
-import com.gomo.app.common.domain.OrderChangeable;
+import com.gomo.app.common.domain.service.DisplayOrder;
+import com.gomo.app.common.domain.service.OrderChangeable;
 
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.AttributeOverrides;
