+++ src/main/java/com/gomo/app/interest/domain/service/ScoreThresholdService.java
@@ -1,6 +1,6 @@
 package com.gomo.app.interest.domain.service;
 
-import com.gomo.app.common.domain.DomainService;
+import com.gomo.app.common.domain.service.DomainService;
 import com.gomo.app.interest.domain.model.Level;
 import com.gomo.app.interest.domain.repository.ScoreThresholdRepository;
