+++ src/main/java/com/gomo/app/point/domain/service/PointService.java
@@ -2,7 +2,7 @@ package com.gomo.app.point.domain.service;
 
 import java.util.List;
 
-import com.gomo.app.common.domain.DomainService;
+import com.gomo.app.common.domain.service.DomainService;
 import com.gomo.app.common.dto.PageRequest;
 import com.gomo.app.member.domain.model.MemberId;
 import com.gomo.app.point.domain.model.Point;
