+++ src/main/java/com/gomo/app/quest/application/OrderUpdateAssignQuestUseCase.java
@@ -3,7 +3,7 @@ package com.gomo.app.quest.application;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.gomo.app.common.application.ApplicationService;
-import com.gomo.app.common.domain.OrderChanger;
+import com.gomo.app.common.domain.service.OrderChanger;
 import com.gomo.app.member.domain.model.MemberId;
 import com.gomo.app.quest.domain.model.QuestType;
 import com.gomo.app.quest.domain.repository.AssignQuestRepository;
