+++ src/main/java/com/gomo/app/quest/domain/model/RepeatQuest.java
@@ -4,8 +4,8 @@ import java.util.UUID;
 
 import com.gomo.app.common.domain.Authorizable;
 import com.gomo.app.common.domain.BaseAudit;
-import com.gomo.app.common.domain.DisplayOrder;
-import com.gomo.app.common.domain.OrderChangeable;
+import com.gomo.app.common.domain.service.DisplayOrder;
+import com.gomo.app.common.domain.service.OrderChangeable;
 import com.gomo.app.interest.domain.model.InterestId;
 
 import jakarta.persistence.AttributeOverride;
