+++ src/main/java/com/gomo/app/survey/presentation/response/ReadSurveyItemResponse.java
@@ -1,6 +1,6 @@
 package com.gomo.app.survey.presentation.response;
 
-import com.gomo.app.common.domain.DisplayOrder;
+import com.gomo.app.common.domain.service.DisplayOrder;
 import com.gomo.app.survey.domain.model.SurveyItemId;
 
 import lombok.Getter;
