+++ src/test/java/com/gomo/app/interest/common/fixture/majorinterest/SecondMajorInterestFixture.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Component;
 
-import com.gomo.app.common.domain.DisplayOrder;
+import com.gomo.app.common.domain.service.DisplayOrder;
 import com.gomo.app.interest.domain.model.InterestId;
 import com.gomo.app.interest.domain.model.MajorInterest;
 import com.gomo.app.interest.domain.model.MajorInterestId;
