+++ src/test/java/com/gomo/app/quest/common/fixture/assign/SpringAssignQuestFixture.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Component;
 
-import com.gomo.app.common.domain.DisplayOrder;
+import com.gomo.app.common.domain.service.DisplayOrder;
 import com.gomo.app.quest.domain.model.AssignQuest;
 import com.gomo.app.quest.domain.model.AssignQuestId;
 import com.gomo.app.quest.domain.model.CompletionProof;
